refactor: Make PagedKeyModel available in cpp
Moved implementation of QmlPagedKeyModel from Qml plugin to library
