+++ extensions/minervous.streamdeck/src/QmlPagedKeyModel.hpp
@@ -3,91 +3,46 @@
 #include <QtQml/QQmlEngine>
 #include <QtCore/qcoreevent.h>
 
-#include "QmlKeyModel.hpp"
+#include "minervous/streamdeck/PagedKeyModel.hpp"
 
 
 namespace minervous::streamdeck
 {
-	class QmlPagedKeyModel: public QmlKeyModel, public QQmlParserStatus
+	class QmlPagedKeyModel: public PagedKeyModel, public QQmlParserStatus
 	{
 		Q_OBJECT
 		QML_ELEMENT
 		Q_INTERFACES(QQmlParserStatus)
 		QML_NAMED_ELEMENT(PagedKeyModel)
-		using Base = minervous::streamdeck::QmlKeyModel;
+		using Base = minervous::streamdeck::PagedKeyModel;
 
-		Q_PROPERTY(QObject * sourceModel READ sourceModel WRITE setSourceModel NOTIFY sourceModelChanged REQUIRED FINAL)
-		Q_PROPERTY(int keysPerPage READ keysPerPage WRITE setKeysPerPage NOTIFY keysPerPageChanged REQUIRED FINAL)
-		Q_PROPERTY(QObject * prevPageKeyEntry READ prevPageKeyEntry WRITE setPrevPageKeyEntry NOTIFY prevPageKeyEntryChanged REQUIRED FINAL)
-		Q_PROPERTY(QObject * nextPageKeyEntry READ nextPageKeyEntry WRITE setNextPageKeyEntry NOTIFY nextPageKeyEntryChanged REQUIRED FINAL)
+		using DefaultPropertyType = QQmlListProperty<QObject>;
+		Q_PROPERTY(DefaultPropertyType data READ qmlData NOTIFY qmlDataChanged)
 
-		Q_PROPERTY(int page READ page WRITE setPage NOTIFY pageChanged FINAL)
-		Q_PROPERTY(int pagesCount READ pagesCount NOTIFY pagesCountChanged FINAL)
+		Q_CLASSINFO("DefaultProperty", "data")
 	public:
 		explicit QmlPagedKeyModel(QObject * parent = nullptr);
 		~QmlPagedKeyModel() override = default;
 
-		QObject * sourceModel() const;
-		void setSourceModel(QObject * model);
-
-		int keysPerPage() const;
-		void setKeysPerPage(int count);
-
-		QObject * prevPageKeyEntry() const;
-		void setPrevPageKeyEntry(QObject * entry);
-
-		QObject * nextPageKeyEntry() const;
-		void setNextPageKeyEntry(QObject * entry);
-
-		int page() const;
-		void setPage(int page);
-
-		int pagesCount() const;
-
 		void classBegin() override;
 		void componentComplete() override;
 
+		Q_INVOKABLE const BaseKeyEntry * at(qsizetype index) const;
+		Q_INVOKABLE BaseKeyEntry * at(qsizetype index);
+
 		Q_INVOKABLE void nextPage();
+
 		Q_INVOKABLE void prevPage();
 
 	signals:
-		void sourceModelChanged();
-		void keysPerPageChanged();
-		void prevPageKeyEntryChanged();
-		void nextPageKeyEntryChanged();
-		void pageChanged();
-		void pagesCountChanged();
-
-	protected:
-		void customEvent(QEvent *event) override;
+		void qmlDataChanged();
 
 	private:
-		class DelayedKeysUpdateEvent: public QEvent
-		{
-		public:
-			DelayedKeysUpdateEvent(QmlPagedKeyModel * sender);
+		DefaultPropertyType qmlData();
 
-			static QEvent::Type staticType();
-
-			QPointer<QmlPagedKeyModel> sender;
-		};
-
-		KeyModel * _sourceModel = nullptr;
-		BaseKeyEntry * _prevPageEntry = nullptr;
-		BaseKeyEntry * _nextPageEntry = nullptr;
-		int _keysPerPage = 0;
-		int _page = 0;
-		int _pagesCount = 0;
-		int _keyIndexBegin = 0;
-		int _keyIndexEnd = 0;
-		bool _completed = false;
-		bool _waitingForLaterUpdateKeys = false;
-
-		void calculatePagesCount();
-		void addOrReplaceEntry(int index, BaseKeyEntry * entry);
-		void updateKeys(bool forceUpdate = false);
-		void updateKeysLater();
-
+		static void qmlAppend(DefaultPropertyType * list, QObject * object);
+		static qsizetype qmlCount(DefaultPropertyType * list);
+		static QObject * qmlAt(DefaultPropertyType * list, qsizetype index);
 	};
 
 }  // namespace minervous::streamdeck
